refactor modules & add FileSagaLogConfig (#187)
Refactor config/modules.go into files per config section. This should make understanding config easier, and various versions of things will be next to each other in one place.

Also added FileSagaLog Config Object, so using this version is configurable.
